+++ src/buffer/channel/topic.rs
@@ -73,6 +73,7 @@ pub fn view<'a>(
             user,
             config.buffer.nickname.show_access_levels,
             config.buffer.nickname.show_bot_icon,
+            false,
             registry,
             &config.display.nickname,
             None,