+++ src/buffer/highlights.rs
@@ -165,6 +165,7 @@ pub fn view<'a>(
                     user,
                     config.buffer.nickname.show_access_levels,
                     config.buffer.nickname.show_bot_icon,
+                    false,
                     registry,
                     &config.display.nickname,
                     config.buffer.nickname.truncate,