+++ src/buffer/input_view.rs
@@ -589,6 +589,7 @@ fn maybe_our_user<'a>(
                     user,
                     config.buffer.text_input.nickname.show_access_levels,
                     config.buffer.nickname.show_bot_icon,
+                    false,
                     registry,
                     &config.display.nickname,
                     None,